+++ .github/workflows/tasks-publish.yml
@@ -50,9 +50,6 @@ jobs:
           git commit . -m "🔖 @huggingface/tasks $BUMPED_VERSION"
           git tag "tasks-v$BUMPED_VERSION"
 
-      - name: "Check Deps are published before publishing this package"
-        run: pnpm -w check-deps gguf
-
       - run: pnpm publish --no-git-checks .
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
