+++ packages/gguf/src/types.ts
@@ -1,5 +1,7 @@
 import type { TransformerLLM } from "./transformer-llm";
 import { LLM_ARCHITECTURES } from "./transformer-llm";
+import type { GGMLQuantizationType } from "@huggingface/tasks";
+export { GGMLQuantizationType } from "@huggingface/tasks";
 
 export type MetadataBaseValue = string | number | bigint | boolean;
 export type MetadataValue = MetadataBaseValue | MetadataBaseValue[] | MetadataValue[]; /// recursive as arrays can be nested.
@@ -43,38 +45,6 @@ export enum GGMLFileQuantizationType {
 	MOSTLY_Q4_0_4_4 = 33,
 	MOSTLY_Q4_0_4_8 = 34,
 	MOSTLY_Q4_0_8_8 = 35,
-}
-
-export enum GGMLQuantizationType {
-	F32 = 0,
-	F16 = 1,
-	Q4_0 = 2,
-	Q4_1 = 3,
-	Q5_0 = 6,
-	Q5_1 = 7,
-	Q8_0 = 8,
-	Q8_1 = 9,
-	Q2_K = 10,
-	Q3_K = 11,
-	Q4_K = 12,
-	Q5_K = 13,
-	Q6_K = 14,
-	Q8_K = 15,
-	IQ2_XXS = 16,
-	IQ2_XS = 17,
-	IQ3_XXS = 18,
-	IQ1_S = 19,
-	IQ4_NL = 20,
-	IQ3_S = 21,
-	IQ2_S = 22,
-	IQ4_XS = 23,
-	I8 = 24,
-	I16 = 25,
-	I32 = 26,
-	I64 = 27,
-	F64 = 28,
-	IQ1_M = 29,
-	BF16 = 30,
 }
 
 export enum GGUFValueType {
